requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.8.2
coverage==7.5.3
cryptography==42.0.5
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jwcrypto==1.5.4
oauthlib==3.2.2
packaging==24.1
pillow==10.2.0
pluggy==1.5.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==1.7.1
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-mock==3.14.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
